+++ NEWS.md
@@ -1,5 +1,7 @@
 # quarto (development version)
 
+- Add `quiet` argument to pass `--quiet` to `quarto` CLI commands
+
 - Error message now advises to re-run with `quiet = FALSE` because `quarto_render(quiet = TRUE)` will run `quarto render` in quiet mode (thanks to @gadenbuie, #126, @wlandau, #16).
 
 - **rsconnect** R package dependency has been moved to Suggest to reduce this package's overall number of dependencies. **rsconnect** package is only required for publishing using `quarto_publish_*()` functions. Users will be prompted to install (when in interactive mode) if not installed.